Rename page graph component
