--- src/app/pages/graph1/graph.component.ts
@@ -1,8 +1,0 @@
-import { Component } from "@angular/core";
-
-@Component({
-    selector: 'app-graph1',
-    templateUrl: './graph1.component.html'
-})
-
-export class Graph1Component{}
